Second design attempt.
The code organization is better.

util: utility code that does not depend on internally defined structures.
  util/shared - `Shared` helps us defer choosing between `Rc` and `Arc`.
model: crate-public static-lifetime models of declared structures
api: public definitions with nice lifetime annotations
  api/project.rs - top-level project encapsulation
  api/shader.rs - top-level shader encapsulation
  api/entry_point.rs - top-level entrypoint encapsulation
  api/data_type - type-mappings from rust types to classes of wgsl types
  api/builder - procedural builder API for constructing shaders
    api/builder/*_builder.rs - the builders passed to builder callbacks
    api/builder/*_handle.rs - lifetime-annotated references to declarations.

The general expectation is that things in `model` share their underlying
contents and are cloneable, and have static lifetime.

Likewise, the top-level `Project` and `Shader` and `EntryPoint` types are
static-lifetime values that share their underlying data and can be cloned.

The types in `api/builder` however, use lifetimes to constrain the mixing
of declarations and enforce scope-compliance.